Extract EngineOptions and find_stracciatella_home
